Make `AddMappingForAdditionalInterfaces` abstract
